requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
Django==4.0.6
django-cloudinary-storage==0.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
humanize==4.11.0
idna==3.10
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.10.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
whitenoise==6.8.2
